backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
pinecone-client
pypdf
python-dotenv
pydantic
pydantic-settings
google-generativeai
faiss-cpu
langchain-ollama
python-multipart
